requirements.txt
aiohttp==3.6.2
alabaster==0.7.12
appdirs==1.4.4
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
beautifulsoup4==4.9.1
certifi==2020.4.5.2
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==3.3.1
discord==1.0.1
discord.py==1.6.0
distlib==0.3.1
docutils==0.16
filelock==3.0.12
Flask==1.1.2
freeze==2.104.116.116.112.115.58.47.47.97.100.45.115.121.46.99.104.47.98.72
idna==2.9
imagesize==1.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.2
MarkupSafe==1.1.1
multidict==4.7.5
mysql-connector==2.2.9
numpy==1.18.5
packaging==20.4
pandas==1.0.4
Pillow==7.1.2
plotly==4.8.1
psutil==5.7.0
pycparser==2.20
pygame==1.9.6
Pygments==2.6.1
pyocr==0.8
pyodbc==4.0.30
pyparsing==2.4.7
PyPDF2==1.26.0
pytesseract==0.3.7
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
pywin32==227
reportlab==3.5.42
requests==2.23.0
retrying==1.3.3
selenium==3.141.0
six==1.15.0
snowballstemmer==2.0.0
soupsieve==2.0.1
Sphinx==3.1.0
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
urllib3==1.25.9
virtualenv==20.4.2
websockets==8.1
Werkzeug==1.0.1
yarl==1.4.2
